# storage/api-docs/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Orders API",
    "description": "API para gestión de órdenes con procesamiento asíncrono",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Servidor local"
    }
  ],
  "paths": {
    "/api/orders/import": {
      "post": {
        "tags": ["Orders"],
        "summary": "Importar órdenes desde archivo CSV o JSON",
        "description": "Permite subir un archivo CSV o JSON con múltiples órdenes para procesarlas de forma asíncrona",
        "operationId": "importOrders",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file", "format"],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Archivo CSV o JSON con las órdenes"
                  },
                  "format": {
                    "type": "string",
                    "enum": ["csv", "json"],
                    "description": "Formato del archivo",
                    "example": "csv"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Órdenes importadas exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "5 órdenes importadas y encoladas para procesamiento"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "total_imported": {
                          "type": "integer",
                          "example": 5
                        },
                        "orders": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Order"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Error de validación",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/orders": {
      "get": {
        "tags": ["Orders"],
        "summary": "Listar todas las órdenes",
        "description": "Obtiene un listado de todas las órdenes con caché Redis",
        "operationId": "listOrders",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Número de página",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "Órdenes por página",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 15
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filtrar por estado",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["pending", "processing", "completed", "failed"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de órdenes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Order"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "current_page": {
                          "type": "integer",
                          "example": 1
                        },
                        "total": {
                          "type": "integer",
                          "example": 50
                        },
                        "per_page": {
                          "type": "integer",
                          "example": 15
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/orders/{id}": {
      "get": {
        "tags": ["Orders"],
        "summary": "Obtener una orden específica",
        "description": "Obtiene los detalles de una orden por su ID",
        "operationId": "getOrder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID de la orden",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalles de la orden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Order"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Orden no encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/orders/{id}/status": {
      "get": {
        "tags": ["Orders"],
        "summary": "Consultar estado de una orden",
        "description": "Obtiene el estado actual de una orden y su historial de cambios",
        "operationId": "getOrderStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID de la orden",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Estado de la orden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "order_id": {
                          "type": "integer",
                          "example": 1
                        },
                        "order_number": {
                          "type": "string",
                          "example": "ORD-001"
                        },
                        "current_status": {
                          "type": "string",
                          "example": "processing"
                        },
                        "history": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/OrderStatusHistory"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Orden no encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "order_number": {
            "type": "string",
            "example": "ORD-001"
          },
          "customer": {
            "type": "string",
            "example": "Juan Pérez"
          },
          "product": {
            "type": "string",
            "example": "Laptop HP"
          },
          "quantity": {
            "type": "integer",
            "example": 2
          },
          "status": {
            "type": "string",
            "enum": ["pending", "processing", "completed", "failed"],
            "example": "processing"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:35:00Z"
          }
        }
      },
      "OrderStatusHistory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "order_id": {
            "type": "integer",
            "example": 1
          },
          "status": {
            "type": "string",
            "example": "processing"
          },
          "notes": {
            "type": "string",
            "example": "Orden enviada al servicio externo"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:35:00Z"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Error en la validación"
          },
          "errors": {
            "type": "object"
          }
        }
      }
    }
  }
}